adad

+++ zari_back/src/main/java/com/ysay/zari_back/repository/StoreRepository.java
@@ -3,5 +3,6 @@ package com.ysay.zari_back.repository;
 import com.ysay.zari_back.entity.Store;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StoreRepository extends JpaRepository<Store, Integer> {
+
+public interface StoreRepository extends JpaRepository<Store, Long> {
 }
